Fix segfault in stack.c
If a destructor is supplied, we only care to free those
elements which remain in the stack. By iterating from
0 to stack->sp, we achieve just that.

Previously, we iterated from 0 to stack->buf_size.
This freed every element that had ever existed in the
stack, even those which had been popped. If the caller of stack_pop
had freed these elements (as they should), then stack_free
would result in a segfault.